\name{sgn}
\alias{sgn}
\title{Sign Function}
\description{
Compute the sign of a numeric value. \\
Returns 1 for positive numbers, -1 for negative numbers, and 0 for zero.
}
\usage{
sgn(x)
}
\arguments{
  \item{x}{a numeric input value or vector.}
}
\value{
A numeric value or vector of 1, -1, or 0 corresponding to the sign of each input element.
}
\examples{
# compute signs of numbers
sgn(3)
sgn(-2)
sgn(0)

}
